requirements.txt
requests
beautifulsoup4
lxml
ollama
python-dotenv
jellyfish